libhsmd: Migrate handle_sign_invoice
